Fix the issue with character detail population

+++ src/components/Character.jsx
@@ -18,6 +18,11 @@ export default function Character({
 
   const deleteCharacter = async (id) => {
     try {
+      const token = localStorage.getItem(TOKEN_NAME);
+      if (!token) {
+        console.log("Token no encontrado");
+        return;
+      }
       await charactersService.delete(id);
       getCharacters();
     } catch (error) {
@@ -29,6 +34,8 @@ export default function Character({
     <div>
       <h1>Name: {name}</h1>
       <Link to={`/characters/${_id}`}>
+        {" "}
+        {/* Envuelve el enlace con la etiqueta Link */}
         <img src={image || defaultImage} alt={name} />
       </Link>
       <p>Race: {race}</p>
@@ -44,11 +51,9 @@ export default function Character({
     </div>
   );
 }
-
 // import React from "react";
 // import { Link } from "react-router-dom";
-// import axios from "axios";
-// import { TOKEN_NAME } from "../context/auth.context";
+// import charactersService from "../services/characters.service";
 
 // export default function Character({
 //   _id,
@@ -60,19 +65,15 @@ export default function Character({
 //   alignment,
 //   image,
 //   getCharacters,
+//   handleClick, // Nueva función de devolución de llamada
 // }) {
 //   const defaultImage =
 //     "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQztsPd6Bijg8DIkFZW_nMaofbIRq_Pm0GR3w&usqp=CAU";
 
 //   const deleteCharacter = async (id) => {
 //     try {
-//       const token = localStorage.getItem(TOKEN_NAME);
-//       await axios.delete(`http://localhost:5005/api/characters/${id}`, {
-//         headers: {
-//           Authorization: `Bearer ${token}`,
-//         },
-//       });
-//       getCharacters(); // Actualiza la lista de personajes después de borrar uno
+//       await charactersService.delete(id);
+//       getCharacters();
 //     } catch (error) {
 //       console.log(error);
 //     }
@@ -81,17 +82,18 @@ export default function Character({
 //   return (
 //     <div>
 //       <h1>Name: {name}</h1>
-//       <img src={image || defaultImage} alt={name} />
+//       <img
+//         src={image || defaultImage}
+//         alt={name}
+//         onClick={() => handleClick(_id)}
+//       />{" "}
+//       {/* Agrega el evento onClick en la imagen */}
 //       <p>Race: {race}</p>
 //       <p>Class: {characterClass}</p>
 //       <p>Level: {level}</p>
 //       <p>Background: {background}</p>
 //       <p>Alignment: {alignment}</p>
-
 //       <button onClick={() => deleteCharacter(_id)}>🗑</button>
-//       <Link to={`/characters/${_id}`}>
-//         <button>View Details</button>
-//       </Link>
 //     </div>
 //   );
 // }
